refactor(web): drop the old client installer
